requirements.txt
Django~=4.1
djangorestframework~=3.13
psycopg2-binary~=2.9
pytest~=6.2
pytest-django~=4.4